Changes to be committed: 	new file:   QQuantLib/AE/ae_classical_qpe.py 	new file:   QQuantLib/AE/ae_iterative_quantum_pe.py 	modified:   QQuantLib/AE/maximum_likelihood_ae.py 	renamed:    QQuantLib/PE/phase_estimation_wqft.py -> QQuantLib/PE/classical_qpe.py 	modified:   QQuantLib/PE/iterative_quantum_pe.py 	modified:   misc/notebooks/04_Classical_Phase_Estimation_Class.ipynb 	modified:   misc/notebooks/05_Iterative_Quantum_Phase_Estimation_Class.ipynb
