+++ QQuantLib/PE/iterative_quantum_pe.py
@@ -406,120 +406,3 @@ class IQPE:
         pds.name = 'Frequency'
         return pd.DataFrame(pds).reset_index()
 
-
-class IQPE_AE:
-    """
-    Class for using Iterative Quantum Phase Estimation (IQPE) class for
-    doing Amplitude Estimation (AE)
-
-    """
-
-    def __init__(self, oracle: qlm.QRoutine, target: list, index: list, **kwargs):
-        """
-
-        Method for initializing the class
-
-        Parameters
-        ----------
-        oracle: QLM gate
-            QLM gate with the Oracle for implementing the
-            Grover operator
-        target : list of ints
-            python list with the target for the amplitude estimation
-        index : list of ints
-            qubits which mark the register to do the amplitude
-            estimation
-
-        kwars : dictionary
-            dictionary that allows the configuration of the IQAE algorithm:
-            Implemented keys:
-        qpu : QLM solver
-            solver for simulating the resulting circutis
-        """
-        #Setting attributes
-        self._oracle = deepcopy(oracle)
-        self._target = check_list_type(target, int)
-        self._index = check_list_type(index, int)
-        #First thing is create the grover operator from the oracle
-        self._grover_oracle = grover(self.oracle,self.target,self.index)
-
-        #Set the QPU to use
-        self.linalg_qpu = kwargs.get('qpu', None)#, get_qpu())
-        if self.linalg_qpu is None:
-            print('Not QPU was provide. Default QPU will be used')
-            self.linalg_qpu = get_default_qpu()
-        self.cbits_number = kwargs.get(
-            'cbits_number', 8)
-        self.shots = kwargs.get('shots', 100)
-
-        #For storing results
-        self.theta = None
-        self.a = None
-        self.iqpe = None
-        self.final_results = None
-    #####################################################################
-    @property
-    def oracle(self):
-        return self._oracle
-
-    @oracle.setter
-    def oracle(self, value):
-        self._oracle = deepcopy(value)
-
-    @property
-    def target(self):
-        return self._target
-
-    @target.setter
-    def target(self, value):
-        self._target = check_list_type(value, int)
-        self._grover_oracle = grover(self.oracle, self.target, self.index)
-
-    @property
-    def index(self):
-        return self._index
-
-    @index.setter
-    def index(self, value):
-        self._index = check_list_type(value, int)
-        self._grover_oracle = grover(self.oracle, self.target, self.index)
-    #####################################################################
-
-    def run(self):
-        r"""
-        run method for the class.
-
-        Parameters
-        ----------
-
-        Returns
-        ----------
-
-        result :
-            the estimation of a
-
-        Notes
-        -----
-        .. math::
-            a  = \cos^2(\theta)
-            \; where \; \theta \; is \;
-            \mathcal{Q}|\Psi\rangle = e^{2i\theta}|\Psi\rangle
-            \; and \; \mathcal{Q} \; the \; Grover \; Operator
-        """
-
-        dict_ae_iqpe = {
-            'initial_state': self.oracle,
-            'unitary_operator': self._grover_oracle,
-            'cbits_number': self.cbits_number,
-            'shots': self.shots,
-            'qpu' : self.linalg_qpu,
-        }
-
-        #Create object IQPE class from Amplitude Estimation inputs
-        self.iqpe_object = IQPE(**dict_ae_iqpe)
-        #Execute IQPE algorithm
-        self.iqpe_object.iqpe()
-        self.final_results = self.iqpe_object.final_results
-        self.theta = self.final_results['theta_90'].iloc[0]
-        self.a = np.cos(self.theta)**2
-        return self.a
